[3D] Replace forced subscription param with shouldSubscribe callback, separate Images and Cameras SceneExtensions
