+++ packages/studio-base/src/panels/ThreeDeeRender/Renderer.ts
@@ -174,10 +174,10 @@ export type RendererSubscription<T = unknown> = {
   preload?: boolean;
   /**
    * By default, topic subscriptions are only created when the topic visibility
-   * has been toggled on by the user in the settings sidebar. Enabling forced
-   * will unconditionally create the topic subscription(s)
+   * has been toggled on by the user in the settings sidebar. Override this
+   * behavior with a custom shouldSubscribe callback.
    */
-  forced?: boolean;
+  shouldSubscribe?: (topic: string) => boolean;
   /** Callback that will be fired for each matching incoming message */
   handler: MessageHandler<T>;
 };
@@ -448,20 +448,20 @@ export class Renderer extends EventEmitter<RendererEvents> {
     // Internal handlers for TF messages to update the transform tree
     this.addSchemaSubscriptions(FRAME_TRANSFORM_DATATYPES, {
       handler: this.handleFrameTransform,
-      forced: true,
+      shouldSubscribe: () => true,
       // Disabled until we can efficiently preload transforms. See
       // <https://github.com/foxglove/studio/issues/4657> for more details.
       // preload: config.scene.transforms?.enablePreloading ?? true,
     });
     this.addSchemaSubscriptions(TF_DATATYPES, {
       handler: this.handleTFMessage,
-      forced: true,
+      shouldSubscribe: () => true,
       // Disabled until we can efficiently preload transforms
       // preload: config.scene.transforms?.enablePreloading ?? true,
     });
     this.addSchemaSubscriptions(TRANSFORM_STAMPED_DATATYPES, {
       handler: this.handleTransformStamped,
-      forced: true,
+      shouldSubscribe: () => true,
       // Disabled until we can efficiently preload transforms
       // preload: config.scene.transforms?.enablePreloading ?? true,
     });
