+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/Cameras.ts
@@ -11,11 +11,10 @@ import type { RosValue } from "@foxglove/studio-base/players/types";
 
 import { BaseUserData, Renderable } from "../Renderable";
 import { Renderer } from "../Renderer";
-import { PartialMessage, PartialMessageEvent, SceneExtension } from "../SceneExtension";
+import { PartialMessageEvent, SceneExtension } from "../SceneExtension";
 import { SettingsTreeEntry } from "../SettingsManager";
 import { makeRgba, rgbaToCssString, stringToRgba } from "../color";
 import { CAMERA_CALIBRATION_DATATYPES } from "../foxglove";
-import { normalizeHeader, normalizeTime } from "../normalizeMessages";
 import {
   CameraInfo,
   CAMERA_INFO_DATATYPES as ROS_CAMERA_INFO_DATATYPES,
@@ -23,9 +22,6 @@ import {
   Marker,
   MarkerAction,
   MarkerType,
-  Matrix3,
-  Matrix3x4,
-  RegionOfInterest,
   TIME_ZERO,
   Vector3,
 } from "../ros";
@@ -33,7 +29,7 @@ import { BaseSettings, fieldLineWidth, PRECISION_DISTANCE } from "../settings";
 import { topicIsConvertibleToSchema } from "../topicIsConvertibleToSchema";
 import { makePose } from "../transforms";
 import { RenderableLineList } from "./markers/RenderableLineList";
-import { projectPixel } from "./projections";
+import { cameraInfosEqual, normalizeCameraInfo, projectPixel } from "./projections";
 
 const log = Logger.getLogger(__filename);
 void log;
@@ -375,99 +371,4 @@ function verticalLine(
     const p = projectPixel(vec3(), uv, cameraModel, settings);
     output.push(p, p);
   }
-}
-
-function cameraInfosEqual(a: CameraInfo | undefined, b: CameraInfo | undefined): boolean {
-  if (!a || !b) {
-    return a === b;
-  } else if (a === b) {
-    return true;
-  }
-
-  if (
-    !(
-      a.header.frame_id === b.header.frame_id &&
-      a.width === b.width &&
-      a.height === b.height &&
-      a.distortion_model === b.distortion_model &&
-      a.binning_x === b.binning_x &&
-      a.binning_y === b.binning_y &&
-      a.roi.x_offset === b.roi.x_offset &&
-      a.roi.y_offset === b.roi.y_offset &&
-      a.roi.height === b.roi.height &&
-      a.roi.width === b.roi.width &&
-      a.roi.do_rectify === b.roi.do_rectify &&
-      a.D.length === b.D.length
-    )
-  ) {
-    return false;
-  }
-  for (let i = 0; i < a.D.length; i++) {
-    if (a.D[i] !== b.D[i]) {
-      return false;
-    }
-  }
-  for (let i = 0; i < 9; i++) {
-    if (a.K[i] !== b.K[i]) {
-      return false;
-    }
-  }
-  for (let i = 0; i < 9; i++) {
-    if (a.R[i] !== b.R[i]) {
-      return false;
-    }
-  }
-  for (let i = 0; i < 12; i++) {
-    if (a.P[i] !== b.P[i]) {
-      return false;
-    }
-  }
-  return true;
-}
-
-function normalizeRegionOfInterest(
-  roi: PartialMessage<RegionOfInterest> | undefined,
-): RegionOfInterest {
-  if (!roi) {
-    return { x_offset: 0, y_offset: 0, height: 0, width: 0, do_rectify: false };
-  }
-  return {
-    x_offset: roi.x_offset ?? 0,
-    y_offset: roi.y_offset ?? 0,
-    height: roi.height ?? 0,
-    width: roi.width ?? 0,
-    do_rectify: roi.do_rectify ?? false,
-  };
-}
-
-function normalizeCameraInfo(
-  message: PartialMessage<IncomingCameraInfo> & PartialMessage<CameraCalibration>,
-): CameraInfo {
-  // Handle lowercase field names as well (ROS2 compatibility)
-  const D = message.D ?? message.d;
-  const K = message.K ?? message.k;
-  const R = message.R ?? message.r;
-  const P = message.P ?? message.p;
-
-  const Dlen = D?.length ?? 0;
-  const Klen = K?.length ?? 0;
-  const Rlen = R?.length ?? 0;
-  const Plen = P?.length ?? 0;
-
-  return {
-    header:
-      "timestamp" in message
-        ? { stamp: normalizeTime(message.timestamp), frame_id: message.frame_id ?? "" }
-        : normalizeHeader(message.header),
-    height: message.height ?? 0,
-    width: message.width ?? 0,
-    distortion_model: message.distortion_model ?? "",
-    D: Dlen > 0 ? (D as number[]) : [],
-    K: Klen === 9 ? (K as Matrix3) : [],
-    R: Rlen === 9 ? (R as Matrix3) : [],
-    P: Plen === 12 ? (P as Matrix3x4) : [],
-    binning_x: message.binning_x ?? 0,
-    binning_y: message.binning_y ?? 0,
-    roi: normalizeRegionOfInterest(message.roi),
-  };
 }
